:art: Rename storybook Text to Typography
