+++ src/stories/Button/index.tsx
@@ -1,6 +1,6 @@
 import { ButtonHTMLAttributes, FC, ReactNode } from "react";
 import { clsx } from "clsx";
-import Text from "../Text/Text";
+import Text from "../Typography/Text";
 
 interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   onClick?: () => void;
