+++ src/stories/Typography/Caption/index.stories.ts
@@ -1,27 +1,27 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import Footnote from ".";
+import Caption from ".";
 
 const meta = {
-  title: "Base/Text/Footnote",
-  component: Footnote,
+  title: "Base/Typography/Caption",
+  component: Caption,
   parameters: {
     layout: "centered",
   },
   tags: ["autodocs"],
-} satisfies Meta<typeof Footnote>;
+} satisfies Meta<typeof Caption>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Sample: Story = {
   args: {
-    children: "Footnote",
+    children: "Caption",
   },
 };
 
 export const Bold: Story = {
   args: {
-    children: "Footnote",
+    children: "Caption",
     bold: true,
   },
 };
