+++ src/stories/Typography/Footnote/index.tsx
