+++ src/stories/Typography/Text/index.stories.ts
@@ -1,27 +1,27 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import Title from ".";
+import Text from ".";
 
 const meta = {
-  title: "Base/Text/Title",
-  component: Title,
+  title: "Base/Typography/Text",
+  component: Text,
   parameters: {
     layout: "centered",
   },
   tags: ["autodocs"],
-} satisfies Meta<typeof Title>;
+} satisfies Meta<typeof Text>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Sample: Story = {
   args: {
-    children: "Title",
+    children: "Text",
   },
 };
 
 export const Bold: Story = {
   args: {
-    children: "Title",
+    children: "Text",
     bold: true,
   },
 };
