+++ src/stories/Typography/Title/index.stories.ts
@@ -1,27 +1,27 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import Headline from ".";
+import Title from ".";
 
 const meta = {
-  title: "Base/Text/Headline",
-  component: Headline,
+  title: "Base/Typography/Title",
+  component: Title,
   parameters: {
     layout: "centered",
   },
   tags: ["autodocs"],
-} satisfies Meta<typeof Headline>;
+} satisfies Meta<typeof Title>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Sample: Story = {
   args: {
-    children: "Headline",
+    children: "Title",
   },
 };
 
 export const Bold: Story = {
   args: {
-    children: "Headline",
+    children: "Title",
     bold: true,
   },
 };
